requirements.txt
python-vagrant==0.5.15
fabric==2.6.0
requests==2.25.1
simplejson==3.17.2
tox==3.22.0
coverage==5.4
PyYAML==5.4.1
straight.plugin==1.5.0
paramiko==2.7.2
pymetasploit3==1.0.3
pylint==2.9.3
flask==2.0.2
pydantic==1.8.2
dotmap==1.3.25

# Sphinx stuff
sphinx-argparse==0.2.5
sphinxcontrib-autoyaml==0.6.1
sphinx-pyreverse==0.0.13
sphinxcontrib.asciinema==0.3.2
# sphinx-pydantic    # This one has issues that must be fixed upstream first

# Mypy stuff
mypy==0.910
types-PyYAML==5.4.6
types-requests==2.25.6
types-simplejson==3.17.0
types-paramiko==2.7.0
